test(integration): factor out helper class
Put helper methods in own class - makes it easier to see what is tests
and allows to split up the tests without too much duplication.
